@@ -3,14 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    
     "@reduxjs/toolkit": "^1.6.1",
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.5.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^7.2.5",
-    "react-scripts": "4.0.1",
+    "react-scripts": "5.0.1",
     "redux": "^4.1.1",
     "redux-thunk": "^2.3.0",
     "web-vitals": "^0.2.4"
